moved components out of 'css' sub-folders
