+++ frontend/src/components/FormContainer.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+// does not seem to be used at the moment - jb
+
 export default function FormContainer({ children }) {
   return (
     <div className="fixed inset-0 dark:bg-primary bg-white -z-10 flex justify-center items-center">
