+++ frontend/src/components/FormTitle.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Title({ children }) {
+export default function FormTitle({ children }) {
   return <h1 className="text-xl text-gray-600 font-semibold text-center">{children}</h1>;
 }
